Move keyword normalization to collect.c as CollectArgs_NormalizeKeywords
Co-locate the keyword list with COLLECT's parser instead of keeping a copy in
generic plan-parsing code; aggregate_request.c keeps only the gated call.
